Move ServiceArtifactId to common: [ECR-4279]
It might be used by some clients (e.g., the Java service launcher)
to parse the SAID.
