+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/runtime/ServiceWrapperIntegrationTest.java
@@ -33,6 +33,7 @@ import com.exonum.binding.common.crypto.CryptoFunctions.Ed25519;
 import com.exonum.binding.common.crypto.PublicKey;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.hash.Hashing;
+import com.exonum.binding.common.runtime.ServiceArtifactId;
 import com.exonum.binding.core.blockchain.BlockchainData;
 import com.exonum.binding.core.service.Configurable;
 import com.exonum.binding.core.service.Configuration;
