+++ exonum-java-binding/integration-tests/src/test/java/com/exonum/binding/test/TestArtifactInfo.java
@@ -18,7 +18,7 @@ package com.exonum.binding.test;
 
 import static com.google.common.base.Preconditions.checkState;
 
-import com.exonum.binding.core.runtime.ServiceArtifactId;
+import com.exonum.binding.common.runtime.ServiceArtifactId;
 import com.google.common.base.Strings;
 import java.nio.file.Path;
 import java.nio.file.Paths;
