+++ exonum-java-binding/site-examples/src/test/java/com/exonum/binding/example/testing/TestkitExamplesTest.java
@@ -28,10 +28,10 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.exonum.binding.common.blockchain.ExecutionStatuses;
 import com.exonum.binding.common.crypto.KeyPair;
 import com.exonum.binding.common.message.TransactionMessage;
+import com.exonum.binding.common.runtime.ServiceArtifactId;
 import com.exonum.binding.core.blockchain.Block;
 import com.exonum.binding.core.blockchain.Blockchain;
 import com.exonum.binding.core.blockchain.BlockchainData;
-import com.exonum.binding.core.runtime.ServiceArtifactId;
 import com.exonum.binding.core.storage.indices.MapIndex;
 import com.exonum.binding.fakeservice.FakeSchema;
 import com.exonum.binding.fakeservice.Transactions.PutTransactionArgs;
